+++ pyccel/ast/builtin_methods/set_methods.py
@@ -10,7 +10,7 @@ always available.
 This module contains objects which describe these methods within Pyccel's AST.
 """
 from pyccel.ast.datatypes import VoidType
-from pyccel.ast.internals import PyccelInternalFunction
+from pyccel.ast.internals import PyccelFunction
 from pyccel.ast.basic import TypedAstNode
 
 __all__ = (
@@ -23,7 +23,7 @@ __all__ = (
     'SetRemove'
 )
 
-class SetMethod(PyccelInternalFunction):
+class SetMethod(PyccelFunction):
     """
     Abstract class for set method calls.