+++ pyccel/ast/numpyext.py
@@ -27,7 +27,7 @@ from .datatypes      import PrimitiveBooleanType, PrimitiveIntegerType, Primitiv
 from .datatypes      import HomogeneousTupleType, FixedSizeNumericType, GenericType, HomogeneousContainerType
 from .datatypes      import InhomogeneousTupleType, ContainerType
 
-from .internals      import PyccelInternalFunction, Slice
+from .internals      import PyccelFunction, Slice
 from .internals      import PyccelArraySize, PyccelArrayShapeElement
 
 from .literals       import LiteralInteger, LiteralString, convert_to_literal
@@ -524,7 +524,7 @@ class NumpyComplex128(NumpyComplex):
 
 #=======================================================================================
 
-class NumpyResultType(PyccelInternalFunction):
+class NumpyResultType(PyccelFunction):
     """
     Class representing a call to the `numpy.result_type` function.
 
@@ -607,7 +607,7 @@ def process_dtype(dtype):
         raise TypeError(f'Unknown type of {dtype}.')
 
 #==============================================================================
-class NumpyNewArray(PyccelInternalFunction):
+class NumpyNewArray(PyccelFunction):
     """
     Superclass for nodes representing NumPy array allocation functions.
 
@@ -617,7 +617,7 @@ class NumpyNewArray(PyccelInternalFunction):
     Parameters
     ----------
     *args : tuple of TypedAstNode
-        The arguments of the superclass PyccelInternalFunction.
+        The arguments of the superclass PyccelFunction.
     class_type : NumpyNDArrayType
         The type of the new array.
     init_dtype : PythonType, PyccelFunctionDef, LiteralString, str
@@ -838,7 +838,7 @@ class NumpyArange(NumpyNewArray):
         return PyccelAdd(self.start, step, simplify=True)
 
 #==============================================================================
-class NumpySum(PyccelInternalFunction):
+class NumpySum(PyccelFunction):
     """
     Represents a call to  numpy.sum for code generation.
 
@@ -869,7 +869,7 @@ class NumpySum(PyccelInternalFunction):
         return self._args[0]
 
 #==============================================================================
-class NumpyProduct(PyccelInternalFunction):
+class NumpyProduct(PyccelFunction):
     """
     Represents a call to numpy.prod for code generation.
 
@@ -905,7 +905,7 @@ class NumpyProduct(PyccelInternalFunction):
 
 
 #==============================================================================
-class NumpyMatmul(PyccelInternalFunction):
+class NumpyMatmul(PyccelFunction):
     """
     Represents a call to numpy.matmul for code generation.
 
@@ -967,7 +967,7 @@ class NumpyMatmul(PyccelInternalFunction):
         return self._args[1]
 
 #==============================================================================
-class NumpyShape(PyccelInternalFunction):
+class NumpyShape(PyccelFunction):
     """
     Represents a call to numpy.shape for code generation.
 
@@ -1132,7 +1132,7 @@ class NumpyLinspace(NumpyNewArray):
         return True
 
 #==============================================================================
-class NumpyWhere(PyccelInternalFunction):
+class NumpyWhere(PyccelFunction):
     """
     Represents a call to `numpy.where`.
 
@@ -1203,7 +1203,7 @@ class NumpyWhere(PyccelInternalFunction):
         return True
 
 #==============================================================================
-class NumpyRand(PyccelInternalFunction):
+class NumpyRand(PyccelFunction):
     """
     Represents a call to  numpy.random.random or numpy.random.rand for code generation.
 
@@ -1228,7 +1228,7 @@ class NumpyRand(PyccelInternalFunction):
             self._class_type = NumpyNDArrayType(NumpyFloat64Type(), rank, order)
 
 #==============================================================================
-class NumpyRandint(PyccelInternalFunction):
+class NumpyRandint(PyccelFunction):
     """
     Class representing a call to NumPy's randint function.
 
@@ -1442,7 +1442,7 @@ class NumpyOnes(NumpyAutoFill):
         return convert_to_literal(1, self.dtype)
 
 #==============================================================================
-class NumpyFullLike(PyccelInternalFunction):
+class NumpyFullLike(PyccelFunction):
     """
     Represents a call to numpy.full_like for code generation.
 
@@ -1489,7 +1489,7 @@ class NumpyFullLike(PyccelInternalFunction):
         return NumpyFull(shape, fill_value, dtype, order)
 
 #==============================================================================
-class NumpyEmptyLike(PyccelInternalFunction):
+class NumpyEmptyLike(PyccelFunction):
     """
     Represents a call to numpy.empty_like for code generation.
 
@@ -1535,7 +1535,7 @@ class NumpyEmptyLike(PyccelInternalFunction):
         return NumpyEmpty(shape, dtype, order)
 
 #==============================================================================
-class NumpyOnesLike(PyccelInternalFunction):
+class NumpyOnesLike(PyccelFunction):
     """
     Represents a call to numpy.ones_like for code generation.
 
@@ -1580,7 +1580,7 @@ class NumpyOnesLike(PyccelInternalFunction):
         return NumpyOnes(shape, dtype, order)
 
 #==============================================================================
-class NumpyZerosLike(PyccelInternalFunction):
+class NumpyZerosLike(PyccelFunction):
     """
     Represents a call to numpy.zeros_like for code generation.
 
@@ -1626,7 +1626,7 @@ class NumpyZerosLike(PyccelInternalFunction):
         return NumpyZeros(shape, dtype, order)
 
 #==============================================================================
-class NumpyNorm(PyccelInternalFunction):
+class NumpyNorm(PyccelFunction):
     """
     Represents call to `numpy.norm`.
 
@@ -1687,7 +1687,7 @@ class NumpyNorm(PyccelInternalFunction):
 # Numpy universal functions
 # https://numpy.org/doc/stable/reference/ufuncs.html#available-ufuncs
 #==============================================================================
-class NumpyUfuncBase(PyccelInternalFunction):
+class NumpyUfuncBase(PyccelFunction):
     """
     Base class for Numpy's universal functions.
 
@@ -2165,7 +2165,7 @@ class NumpyMod(NumpyUfuncBinary):
                 arg_dtype = arg_class_type
             return process_dtype(arg_dtype)
 
-class NumpyAmin(PyccelInternalFunction):
+class NumpyAmin(PyccelFunction):
     """
     Represents a call to  numpy.min for code generation.
 
@@ -2193,7 +2193,7 @@ class NumpyAmin(PyccelInternalFunction):
         """
         return self._args[0]
 
-class NumpyAmax(PyccelInternalFunction):
+class NumpyAmax(PyccelFunction):
     """
     Represents a call to  numpy.max for code generation.
 
@@ -2400,7 +2400,7 @@ class NumpyNonZeroElement(NumpyNewArray):
         """
         return self._dim
 
-class NumpyNonZero(PyccelInternalFunction):
+class NumpyNonZero(PyccelFunction):
     """
     Class representing a call to the function `numpy.nonzero`.
 
@@ -2446,7 +2446,7 @@ class NumpyNonZero(PyccelInternalFunction):
     def __iter__(self):
         return self._elements.__iter__()
 
-class NumpyCountNonZero(PyccelInternalFunction):
+class NumpyCountNonZero(PyccelFunction):
     """
     Class representing a call to the NumPy function `count_nonzero`.
 
@@ -2522,7 +2522,7 @@ class NumpyCountNonZero(PyccelInternalFunction):
         return self._keep_dims
 
 
-class NumpySize(PyccelInternalFunction):
+class NumpySize(PyccelFunction):
     """
     Represent a call to numpy.size in the user code.