+++ pyccel/ast/sysext.py
@@ -6,7 +6,7 @@
 """ Module containing objects from the sys module understood by pyccel
 """
 from .core import PyccelFunctionDef, Module
-from .internals import PyccelInternalFunction
+from .internals import PyccelFunction
 from .datatypes import VoidType
 from .internals import LiteralInteger
 
@@ -17,7 +17,7 @@ __all__ = (
     'sys_mod',
 )
 
-class SysExit(PyccelInternalFunction):
+class SysExit(PyccelFunction):
     """
     Represents a call to sys.exit.